test: remove message from strictEqual assertions
If values are not equals in strictEqual assertions, an AssertionError is
thrown with a message property set equals to the value of the message
parameter. If we pass a message, this message will be printed instead of
the default message, which contains the value that is causing the error.
Hence removed the value passed as the message in strictEqual assertions
of test/parallel/test-http-response-readable.js

PR-URL: https://github.com/nodejs/node/pull/20067
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: Ruben Bridgewater <ruben@bridgewater.de>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Richard Lau <riclau@uk.ibm.com>
Reviewed-By: Rich Trott <rtrott@gmail.com>
Reviewed-By: Trivikram Kamat <trivikr.dev@gmail.com>