doc: remove link prediction from STYLE_GUIDE.md
Remove promise to eventually link an example document to show level-one
heading. The style guide has been fine without the example doc for a
long time. The statement is already clear. And the style guide itself is
an example. We don't link to examples of wrapping at 80 characters, for
example. There's no need to link just to show what a level-one heading
is. (And if level-one heading is unclear, then it may be preferable to
improve the terminology rather than to link to an example.)

PR-URL: https://github.com/nodejs/node/pull/21031
Reviewed-By: Richard Lau <riclau@uk.ibm.com>
Reviewed-By: Vse Mozhet Byt <vsemozhetbyt@gmail.com>
Reviewed-By: Tiancheng "Timothy" Gu <timothygu99@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Trivikram Kamat <trivikr.dev@gmail.com>